Introduced new component to handle drawing of markers on the map
